move extensions guards add resolvers
