+++ npm/ng-packs/packages/account/src/lib/guards/extensions.guard.ts
@@ -18,7 +18,7 @@ import {
 import { eAccountComponents } from '../enums/components';
 
 /**
- * @deprecated Use `AccountExtensionsGuardFn` instead.
+ * @deprecated Use `AccountExtensionsResolver` instead.
  */
 @Injectable()
 export class AccountExtensionsGuard implements IAbpGuard {
@@ -47,28 +47,3 @@ export class AccountExtensionsGuard implements IAbpGuard {
     );
   }
 }
-
-export const AccountExtensionsGuardFn = () => {
-  const configState = inject(ConfigStateService);
-  const extensions = inject(ExtensionsService);
-
-  const config = { optional: true };
-
-  const editFormContributors = inject(ACCOUNT_EDIT_FORM_PROP_CONTRIBUTORS, config) || {};
-
-  return getObjectExtensionEntitiesFromStore(configState, 'Identity').pipe(
-    map(entities => ({
-      [eAccountComponents.PersonalSettings]: entities.User,
-    })),
-    mapEntitiesToContributors(configState, 'AbpIdentity'),
-    tap(objectExtensionContributors => {
-      mergeWithDefaultProps(
-        extensions.editFormProps,
-        DEFAULT_ACCOUNT_FORM_PROPS,
-        objectExtensionContributors.editForm,
-        editFormContributors,
-      );
-    }),
-    map(() => true),
-  );
-};
