+++ npm/ng-packs/packages/account/src/lib/resolvers/extensions.resolver.ts
@@ -1,0 +1,31 @@
+import { ConfigStateService } from "@abp/ng.core";
+import { ExtensionsService, getObjectExtensionEntitiesFromStore, mapEntitiesToContributors, mergeWithDefaultProps } from "@abp/ng.theme.shared/extensions";
+import { inject } from "@angular/core";
+import { map, tap } from "rxjs";
+import { eAccountComponents } from "../enums";
+import { ACCOUNT_EDIT_FORM_PROP_CONTRIBUTORS, DEFAULT_ACCOUNT_FORM_PROPS } from "../tokens";
+
+export const AccountExtensionsResolver = () => {
+  const configState = inject(ConfigStateService);
+  const extensions = inject(ExtensionsService);
+
+  const config = { optional: true };
+
+  const editFormContributors = inject(ACCOUNT_EDIT_FORM_PROP_CONTRIBUTORS, config) || {};
+
+  return getObjectExtensionEntitiesFromStore(configState, 'Identity').pipe(
+    map(entities => ({
+      [eAccountComponents.PersonalSettings]: entities.User,
+    })),
+    mapEntitiesToContributors(configState, 'AbpIdentity'),
+    tap(objectExtensionContributors => {
+      mergeWithDefaultProps(
+        extensions.editFormProps,
+        DEFAULT_ACCOUNT_FORM_PROPS,
+        objectExtensionContributors.editForm,
+        editFormContributors,
+      );
+    }),
+    map(() => true),
+  );
+};
