+++ npm/ng-packs/packages/account/src/lib/resolvers/index.ts
@@ -1,0 +1,1 @@
+export * from './extensions.resolver'
