+++ npm/ng-packs/packages/tenant-management/src/lib/resolvers/index.ts
@@ -1,0 +1,1 @@
+export * from './extensions.resolver'
